refactor to allow reuse of exe utility code
